requirements.txt
Django == 1.10.5
channels == 1.0.2
django_countries == 4.0
django_registration == 2.2
nltk == 3.2.1
numpy == 1.11.2
scipy == 0.18.1
pandas == 0.19.1
scikit_learn == 0.18
whitenoise == 3.3.0
django-crispy-forms == 1.6.1
psycopg2 == 2.6.2
waitress == 1.0.2
dj-database-url == 0.4.2
pillow == 3.4.2
asgi_redis == 1.0.0